+++ packages/plugins/keyboard/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @splicetree/plugin-keyboard
 
+## 2.0.0
+
 ## 1.1.0
 
 ## 1.0.0